requirements.txt
fire>=0.3.1
pyyaml